refactor(core): extract interaction and loading UI
